+++ cnccode/customClasses/ChargeableWorkCustomerRequest/Core/ChargeableWorkCustomerRequestRepository.php
@@ -1,0 +1,7 @@
+<?php
+
+namespace CNCLTD\ChargeableWorkCustomerRequest\Core;
+interface ChargeableWorkCustomerRequestRepository
+{
+    public function getNextIdentity(): ChargeableWorkCustomerRequestId;
+}